+++ Release_Notes.md
@@ -5,3 +5,7 @@
 - Added `lunarphase` module successfully.
 - Testing new `pokemon` module
 - Testing declaration of widgets in wtfable
+
+## Pokemon Go API
+
+- Testing use of pokemon-go API